Refactor into individual parameter resolver for better understanding
