+++ src/main/java/io/github/uchagani/jp/InjectPlaywright.java
@@ -7,7 +7,8 @@ import java.lang.annotation.*;
 @Inherited
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.METHOD})
-@ExtendWith(PlaywrightExtension.class)
+@ExtendWith({PlaywrightParameterResolver.class, BrowserParameterResolver.class, BrowserContextParameterResolver.class,
+             PageParameterResolver.class, PlaywrightTestWatcher.class})
 public @interface InjectPlaywright {
     Class<? extends PlaywrightConfig> value() default PlaywrightConfig.class;
 }
